backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
sentence-transformers
faiss-cpu
langdetect
python-dotenv
aiohttp
